reestruturando modelos do projeto

+++ Escon.SisctNET.Model/Access.cs
@@ -10,14 +10,14 @@ namespace Escon.SisctNET.Model
         [Required(ErrorMessage = "Obrigatório!")]
         [Display(Name = "Funcionalidade")]
         [ForeignKey("Functionality")]
-        public Nullable<int> FunctionalityId { get; set; }
+        public Nullable<long> FunctionalityId { get; set; }
 
         public virtual Functionality Functionality { get; set; }
 
         [Required(ErrorMessage = "Obrigatório!")]
         [Display(Name = "Pefil")]
         [ForeignKey("Profile")]
-        public Nullable<int> ProfileId { get; set; }
+        public Nullable<long> ProfileId { get; set; }
 
         public virtual Profile Profile { get; set; }
