+++ Escon.SisctNET.Model/CompraAnexo.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.EntityFrameworkCore.Infrastructure;
-using System;
 using System.ComponentModel.DataAnnotations.Schema;
 
 namespace Escon.SisctNET.Model
@@ -11,7 +10,7 @@ namespace Escon.SisctNET.Model
         public ILazyLoader LazyLoader { get; set; }
 
         [ForeignKey("TaxAnexo")]
-        public int TaxAnexoId { get; set; }
+        public long TaxAnexoId { get; set; }
 
         private TaxAnexo taxAnexo;
