+++ Escon.SisctNET.Model/Note.cs
@@ -9,7 +9,7 @@ namespace Escon.SisctNET.Model
     {
         [Display(Name = "Empresa")]
         [ForeignKey("Company")]
-        public Nullable<int> CompanyId { get; set; }
+        public Nullable<long> CompanyId { get; set; }
 
         public virtual Company Company { get; set; }
