+++ Escon.SisctNET.Model/NoteBase.cs
@@ -6,9 +6,10 @@ namespace Escon.SisctNET.Model
 {
     public class NoteBase : EntityBase
     {
+
         [Display(Name = "Empresa")]
         [ForeignKey("Company")]
-        public Nullable<int> CompanyId { get; set; }
+        public Nullable<long> CompanyId { get; set; }
 
         public virtual Company Company { get; set; }
