+++ Escon.SisctNET.Repository/ICreditBalanceRepository.cs
@@ -4,8 +4,8 @@ namespace Escon.SisctNET.Repository
 {
     public interface ICreditBalanceRepository : IRepository<CreditBalance>
     {
-        CreditBalance FindByLastMonth(int companyid, string month, string year, Log log = null);
+        CreditBalance FindByLastMonth(long companyid, string month, string year, Log log = null);
 
-        CreditBalance FindByCurrentMonth(int companyid, string month, string year, Log log = null);
+        CreditBalance FindByCurrentMonth(long companyid, string month, string year, Log log = null);
     }
 }
