+++ Escon.SisctNET.Repository/IEstoqueRepository.cs
@@ -4,6 +4,6 @@ namespace Escon.SisctNET.Repository
 {
     public interface IEstoqueRepository : IRepository<Model.Estoque>
     {
-        List<Model.Estoque> FindByCompany(int company, Model.Log log = null);
+        List<Model.Estoque> FindByCompany(long company, Model.Log log = null);
     }
 }
