+++ Escon.SisctNET.Repository/ILogRepository.cs
@@ -6,8 +6,8 @@ namespace Escon.SisctNET.Repository
     {
         List<Model.Log> FindAll(Model.Log log);
 
-        List<Model.Log> FindUser(int userId, Model.Log log);
+        List<Model.Log> FindUser(long userId, Model.Log log);
 
-        List<Model.Log> FindFunctionality(int functionalityId, Model.Log log);
+        List<Model.Log> FindFunctionality(long functionalityId, Model.Log log);
     }
 }
