+++ Escon.SisctNET.Repository/IProduct2Repository.cs
@@ -11,11 +11,11 @@ namespace Escon.SisctNET.Repository
 
         Model.Product2 FindByDescription(string description, Model.Log log = null);
 
-        decimal FindByPrice(int id, Model.Log log = null);
+        decimal FindByPrice(long id, Model.Log log = null);
 
-        Model.Product2 FindByProduct(string code, int grupoId, Model.Log log = null);
+        Model.Product2 FindByProduct(string code, long grupoId, Model.Log log = null);
 
-        List<Model.Product2> FindByGroup(int groupid, Model.Log log = null);
+        List<Model.Product2> FindByGroup(long groupid, Model.Log log = null);
 
         List<Model.Product2> FindAllInDate2(DateTime dateProd, Model.Log log = null);
     }
