+++ Escon.SisctNET.Repository/IProductNoteInventoryEntryRepository.cs
@@ -8,14 +8,14 @@ namespace Escon.SisctNET.Repository
 
         void Create(List<Model.ProductNoteInventoryEntry> products, Model.Log log = null);
 
-        List<Model.ProductNoteInventoryEntry> FindByCompany(int companyId, Model.Log log = null);
+        List<Model.ProductNoteInventoryEntry> FindByCompany(long companyId, Model.Log log = null);
 
-        List<Model.ProductNoteInventoryEntry> FindByNotes(int companyId, string year, string month, Model.Log log = null);
+        List<Model.ProductNoteInventoryEntry> FindByNotes(long companyId, string year, string month, Model.Log log = null);
 
         List<Model.ProductNoteInventoryEntry> FindByNote(string chave, Model.Log log = null);
 
-        List<Model.ProductNoteInventoryEntry> FindByPeriod(int companyId, DateTime inicio, DateTime fim, Model.Log log = null);
+        List<Model.ProductNoteInventoryEntry> FindByPeriod(long companyId, DateTime inicio, DateTime fim, Model.Log log = null);
 
-        List<Model.ProductNoteInventoryEntry> FindByNotes(int companyId, string year, Model.Log log = null);
+        List<Model.ProductNoteInventoryEntry> FindByNotes(long companyId, string year, Model.Log log = null);
     }
 }
