+++ Escon.SisctNET.Repository/IProductNoteInventoryExitRepository.cs
@@ -8,14 +8,14 @@ namespace Escon.SisctNET.Repository
 
         void Create(List<Model.ProductNoteInventoryExit> products, Model.Log log = null);
 
-        List<Model.ProductNoteInventoryExit> FindByCompany(int companyId, Model.Log log = null);
+        List<Model.ProductNoteInventoryExit> FindByCompany(long companyId, Model.Log log = null);
 
-        List<Model.ProductNoteInventoryExit> FindByNotes(int companyId, string year, string month, Model.Log log = null);
+        List<Model.ProductNoteInventoryExit> FindByNotes(long companyId, string year, string month, Model.Log log = null);
 
         List<Model.ProductNoteInventoryExit> FindByNote(string chave , Model.Log log = null);
 
-        List<Model.ProductNoteInventoryExit> FindByPeriod(int companyId, DateTime inicio, DateTime fim, Model.Log log = null);
+        List<Model.ProductNoteInventoryExit> FindByPeriod(long companyId, DateTime inicio, DateTime fim, Model.Log log = null);
 
-        List<Model.ProductNoteInventoryExit> FindByNotes(int companyId, string year, Model.Log log = null);
+        List<Model.ProductNoteInventoryExit> FindByNotes(long companyId, string year, Model.Log log = null);
     }
 }
