+++ Escon.SisctNET.Repository/IProductNoteRepository.cs
@@ -11,7 +11,7 @@ namespace Escon.SisctNET.Repository
 
         void Update(List<Model.ProductNote> products, Model.Log log = null);
 
-        List<Model.ProductNote> FindByNote(int noteId, Model.Log log = null);
+        List<Model.ProductNote> FindByNote(long noteId, Model.Log log = null);
 
         List<Model.ProductNote> FindByProducts(List<Model.Note> notes, Model.Log log = null);
 
@@ -21,24 +21,24 @@ namespace Escon.SisctNET.Repository
 
         List<Model.ProductNote> FindByCnpjCprod(List<Model.Note> notes, string cnpj, string cprod, string ncm, string cest, Model.Log log = null);
 
-        decimal FindByTotal(List<int> notes, Model.Log log = null);
+        decimal FindByTotal(List<long> notes, Model.Log log = null);
 
         decimal FindBySubscription(List<Model.Note> notes, Model.TypeTaxation taxationType, Model.Log log = null);
 
-        List<Model.ProductNote> FindByTaxation(int noteId, Model.Log log = null);
+        List<Model.ProductNote> FindByTaxation(long noteId, Model.Log log = null);
 
-        Model.ProductNote FindByProduct(int noteId, string item, Model.Log log = null);
+        Model.ProductNote FindByProduct(long noteId, string item, Model.Log log = null);
 
-        List<Model.ProductNote> FindByCfopNotesIn(int companyId, List<Model.Note> notes, Model.Log log = null);
+        List<Model.ProductNote> FindByCfopNotesIn(long companyId, List<Model.Note> notes, Model.Log log = null);
 
-        List<Model.ProductNote> FindByCfopNotesOut(int companyId, List<Model.Note> notes, Model.Log log = null);
+        List<Model.ProductNote> FindByCfopNotesOut(long companyId, List<Model.Note> notes, Model.Log log = null);
 
-        bool FindByNcmAnnex(int Annex, string ncm, Model.Log log = null);
+        bool FindByNcmAnnex(long Annex, string ncm, Model.Log log = null);
 
         List<Model.ProductNote> FindByIncentive(List<Model.Note> notes, Model.Log log = null);
 
         List<Model.ProductNote> FindByNormal(List<Model.Note> notes, Model.Log log = null);
 
-        Model.ProductNote FindByProduct(int id, Model.Log log = null);
+        Model.ProductNote FindByProduct(long id, Model.Log log = null);
     }
 }
