+++ Escon.SisctNET.Repository/ITaxProducerRepository.cs
@@ -5,7 +5,7 @@ namespace Escon.SisctNET.Repository
 {
     public interface ITaxProducerRepository : IRepository<Model.TaxProducer>
     {
-        List<Model.TaxProducer> FindByTaxs(int companyid, string month, string year, Model.Log log = null);
+        List<Model.TaxProducer> FindByTaxs(long companyid, string month, string year, Model.Log log = null);
 
         void Create(List<Model.TaxProducer> taxProducers, Model.Log log = null);
