+++ Escon.SisctNET.Repository/ITaxationRepository.cs
@@ -8,7 +8,7 @@ namespace Escon.SisctNET.Repository
     {
         Model.Taxation FindByCode(string code, string cest, DateTime data, Model.Log log = null);
 
-        List<Model.Taxation> FindByCompany(int companyId, Model.Log log = null);
+        List<Model.Taxation> FindByCompany(long companyId, Model.Log log = null);
 
         Model.Taxation FindByNcm(string code, string cest, Model.Log log = null);
