+++ Escon.SisctNET.Repository/IVendaAnexoRepository.cs
@@ -8,6 +8,6 @@ namespace Escon.SisctNET.Repository
 
         void Update(List<Model.VendaAnexo> vendaAnexos, Model.Log log = null);
 
-        List<Model.VendaAnexo> FindByVendasTax(int taxAnexo, Model.Log log = null);
+        List<Model.VendaAnexo> FindByVendasTax(long taxAnexo, Model.Log log = null);
     }
 }
